Add a temporary tool for testing utility functions. Fix function for study area to actually use the mask properties instead of assuming env CellSize and output coordinate system to be the same as the mask's.
