+++ Toolbox/ArcSDM.TestTool.pyt.xml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<metadata xml:lang="fi"><Esri><CreaDate>20241213</CreaDate><CreaTime>16081300</CreaTime><ArcGISFormat>1.0</ArcGISFormat><SyncOnce>TRUE</SyncOnce></Esri></metadata>
