+++ Toolbox/arcsdm/calculateresponse.py
@@ -34,6 +34,17 @@ def Execute(self, parameters, messages):
         # TODO: Add check for weights table columns depending on weights type (unique weights shouldn't have generalized columns)
 
         check_input_data(evidence_rasters, training_point_feature)
+
+        # TODO: check that all evidence rasters have the same cell size?
+        # TODO: use the env Cell Size instead. not all of the evidence rasters necessarily have the same cell size
+        # TODO: evidence rasters should be resampled to env Cell Size?
+        evidence_cellsize = arcpy.Describe(evidence_rasters[0]).MeanCellWidth
+        
+        # TODO: apply mask
+        study_area_sq_km = 0.0
+        study_area_cell_count = 0.0
+
+
         
     except arcpy.ExecuteError:
         arcpy.AddError(arcpy.GetMessages(2))
