+++ Toolbox/arcsdm/calculateweights.py
@@ -6,7 +6,7 @@ import traceback
 
 from arcsdm.common import log_arcsdm_details
 from arcsdm.sdmvalues import log_wofe
-from arcsdm.wofe_common import check_input_data, get_training_point_statistics
+from arcsdm.wofe_common import check_input_data, get_study_area_size_sq_km, get_training_point_statistics
 
 ASCENDING = "Ascending"
 DESCENDING = "Descending"
