+++ Toolbox/arcsdm/sdmvalues.py
@@ -84,70 +84,6 @@ def getPriorProb(TrainPts, unitCell, mapUnits):
     return priorprob
 
 
-def get_mask_area_in_km(mapUnits):
-    """
-    Return the mask size in square kilometers.
-    """
-    try:
-        desc = arcpy.Describe(arcpy.env.mask)
-
-        arcpy.AddMessage(f"mask datatype: {desc.dataType}")
-
-        if desc.dataType in ["RasterLayer", "RasterBand", "RasterDataset"]:
-            if not str(arcpy.env.cellSize).replace('.', '', 1).replace(',', '', 1).isdigit():
-                arcpy.AddMessage("*" * 50)
-                arcpy.AddError("ERROR: Cell Size must be numeric when mask is raster. Check Environments!")
-                arcpy.AddMessage("*" * 50)
-                raise SDMError
-
-            arcpy.AddMessage("Counting raster size")
-            arcpy.AddMessage("File: " + desc.catalogPath)
-            rows = int(arcpy.GetRasterProperties_management(desc.catalogPath, "ROWCOUNT").getOutput(0))
-            columns = int(arcpy.GetRasterProperties_management(desc.catalogPath, "COLUMNCOUNT").getOutput(0))
-            raster_array = arcpy.RasterToNumPyArray(desc.catalogPath, nodata_to_value=-9999)
-            area_sq_m = 0
-            count = 0
-            arcpy.AddMessage("Iterating through mask in numpy..." + str(columns) + "x" + str(rows))
-            for i in range(rows):
-                for j in range(columns):
-                    if raster_array[i][j] != -9999:
-                        count += 1
-            arcpy.AddMessage("count:" + str(count))
-            cellsize = float(str(arcpy.env.cellSize).replace(",", "."))
-            area_sq_m = count * (cellsize * cellsize)
-
-        elif desc.dataType in ["FeatureLayer", "FeatureClass", "ShapeFile"]:
-            maskrows = arcpy.SearchCursor(desc.catalogPath)
-            shapeName = desc.shapeFieldName
-            maskrow = maskrows.next()
-            area_sq_m = 0
-            while maskrow:
-                feat = maskrow.getValue(shapeName)
-                area_sq_m += feat.area
-                maskrow = maskrows.next()
-            arcpy.AddMessage("count:" + str(area_sq_m))
-
-        else:
-            raise arcpy.ExecuteError(desc.dataType + " is not allowed as Mask!")
-
-        mapUnits = mapUnits.lower().strip()
-        if not mapUnits.startswith('meter'):
-            arcpy.AddError('Incorrect output map units: Check units of study area.')
-        conversion = getMapConversion(mapUnits)
-        unit_cell_count = area_sq_m * conversion
-        return unit_cell_count
-    except arcpy.ExecuteError as e:
-        raise
-    except:
-        tb = sys.exc_info()[2]
-        tbinfo = traceback.format_tb(tb)[0]
-        arcpy.AddError(tbinfo)
-        if len(arcpy.GetMessages(2)) > 0:
-            msgs = "SDM GP ERRORS:\n" + arcpy.GetMessages(2) + "\n"
-            arcpy.AddError(msgs)
-        raise
-
-
 # Old
 def getMaskSize(mapUnits):
     """
@@ -213,83 +149,6 @@ def getMaskSize(mapUnits):
             arcpy.AddError(msgs)
         raise
 
-
-def log_wofe(unit_cell_area_sq_km, TrainPts):
-    """
-    Log WofE parameters to Geoprocessor History.
-    """
-    try:
-        arcpy.AddMessage("\n" + "=" * 10 + " WofE parameters " + "=" * 10)
-
-        mapUnits = getMapUnits()
-        mapUnits = mapUnits.lower().strip()
-        if not mapUnits.startswith("meter"):
-            arcpy.AddError("Output map unit should be meter: Check output coordinate system & units of study area.")
-
-        arcpy.AddMessage("%-20s %s" % ("Map Units:", mapUnits))
-
-        total_area = get_mask_area_in_km(mapUnits)
-
-        if not arcpy.env.mask:
-            arcpy.AddError("Study Area mask not set. Check Environments!")
-        else:
-            if not arcpy.Exists(arcpy.env.mask):
-                arcpy.AddError(f"Mask {arcpy.env.mask} not found!")
-            
-            mask_descr = arcpy.Describe(arcpy.env.mask)
-            arcpy.AddMessage("%-20s %s" % ("Mask:", "\"" + mask_descr.name + "\" and it is of type " + mask_descr.dataType))
-            
-            if mask_descr.dataType in ["FeatureLayer", "FeatureClass"]:
-                arcpy.AddWarning("Warning: You should only use single value raster type masks!")
-            
-            arcpy.AddMessage("%-20s %s" % ("Mask size (km^2):", str(total_area)))
-
-        if not arcpy.env.cellSize:
-            arcpy.AddError("Study Area cellsize not set")
-        if arcpy.env.cellSize == "MAXOF":
-            arcpy.AddWarning("Cellsize should have definitive value?")
-
-        cellsize = arcpy.env.cellSize
-        arcpy.AddMessage("%-20s %s" % ("Cell Size:", cellsize))
-        
-        unit_cell_area_sq_km = float(unit_cell_area_sq_km)
-        unit_cells_count = float(total_area) / unit_cell_area_sq_km
-
-        # Note! GetCount does not care about the mask. Ie., it's assumed that
-        # the mask has been already applied at training sites reduction.
-        training_point_count = arcpy.management.GetCount(TrainPts)
-        arcpy.AddMessage("%-20s %s" % ("# Training Sites:", training_point_count))
-        arcpy.AddMessage("%-20s %s" % ("Unit Cell Area:", "{} km^2, Cells in area: {} ".format(unit_cell_area_sq_km, unit_cells_count)))
-
-        if unit_cells_count == 0:
-            arcpy.AddError("ERROR: 0 Cells in Area!")
-
-        priorprob = float(str(training_point_count)) / float(unit_cells_count)
-
-        if not (0 < priorprob <= 1.0):
-            arcpy.AddError('Incorrect no. of training sites or unit cell area. TrainingPointsResult {}'.format(priorprob))
-
-        arcpy.AddMessage("%-20s %0.6f" % ("Prior Probability:", priorprob))
-        arcpy.AddMessage("%-20s %s" % ("Training Points:", arcpy.Describe(TrainPts).catalogPath))
-        arcpy.AddMessage("%-20s %s" % ("Study Area Raster:", arcpy.Describe(arcpy.env.mask).catalogPath))
-        arcpy.AddMessage("%-20s %s" % ("Study Area Area:", str(total_area) + " km^2"))
-        arcpy.AddMessage("")
-    except arcpy.ExecuteError as e:
-        if not all(e.args):
-            arcpy.AddMessage("Calculate weights caught arcpy.ExecuteError: ")
-            args = e.args[0]
-            args.split('\n')
-            arcpy.AddError(args)
-        arcpy.AddMessage("-------------- END EXECUTION ---------------")
-        raise
-    except:
-        tb = sys.exc_info()[2]
-        tbinfo = traceback.format_tb(tb)[0]
-        arcpy.AddError(tbinfo)
-        if len(arcpy.GetMessages(2)) > 0:
-            msgs = "SDM GP ERRORS:\n" + arcpy.GetMessages(2) + "\n"
-            arcpy.AddError(msgs)
-        raise
 
 # Old
 def appendSDMValues(unitCell, TrainPts):
